% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scholarsearch.R
\name{scholar_get}
\alias{scholar_get}
\title{Search google scholar}
\usage{
scholar_get(
  query = NULL,
  as_ylo = NULL,
  as_yhi = NULL,
  source = NULL,
  open_browser = FALSE,
  u = "https://scholar.google.co.uk/scholar"
)
}
\arguments{
\item{query}{Search query}

\item{as_ylo}{Lowest year}

\item{as_yhi}{Highest year}

\item{source}{Journal}

\item{open_browser}{Open the browser?}

\item{u}{Base URL}
}
\description{
Search google scholar
}
\examples{
scholar_get(query = '"open+source"+"transport+planning"')
}
