% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_citations.R
\name{scholar_get_citations}
\alias{scholar_get_citations}
\title{Scrape title, author and number of citations from google scholar url}
\usage{
scholar_get_citations(u)
}
\arguments{
\item{u}{url}
}
\description{
Scrape title, author and number of citations from google scholar url
}
\examples{
query = "open source software"
u = scholarurl(query)
res = scholar_get_citations(u)
res
}
