% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scholarsearch.R
\name{scholarsearch}
\alias{scholarsearch}
\title{Search google scholar}
\usage{
scholarsearch(
  query = NULL,
  as_ylo = NULL,
  as_yhi = NULL,
  source = NULL,
  open_browser = TRUE,
  u = "https://scholar.google.co.uk/scholar"
)
}
\arguments{
\item{query}{Search query}

\item{as_ylo}{Lowest year}

\item{as_yhi}{Highest year}

\item{source}{Journal}

\item{open_browser}{Open the browser?}

\item{u}{Base URL}
}
\description{
Search google scholar
}
\examples{
scholarsearch(query = "bikeshare data analysis")
scholarsearch(query = "bikeshare data analysis", source = "transportation research part A")
scholarsearch(query = "bikeshare data analysis", source = "transportation research part A", as_ylo = 2018)
scholarsearch(query = "bikeshare data analysis", source = "transportation research part A", as_yhi = 2012)
scholarsearch(query = '"open+source"+"transport+planning"', as_yhi = 2012) # note google's citation format
scholarsearch(query = '"open+source"+"transport+planning"', as_yhi = 2012) # note google's citation format
}
