% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scholarsearch.R
\name{scholarurl}
\alias{scholarurl}
\alias{scholarsearch}
\title{Get url of google scholar search}
\usage{
scholarurl(query)
}
\arguments{
\item{query}{Search query}
}
\description{
Get url of google scholar search
}
\examples{
scholarurl("test")
}
